requirements.txt
python-telegram-bot==20.7
APScheduler==3.10.4
aiohttp==3.9.3
aiosqlite==0.20.0
python-dotenv==1.0.1

# Web dashboard
fastapi==0.111.0
uvicorn[standard]==0.29.0
jinja2==3.1.4
python-multipart==0.0.9
itsdangerous==2.2.0
bcrypt==4.1.3
slowapi==0.1.9
